% Auto-generated: do not edit by hand
\name{dashReactGrid}

\alias{dashReactGrid}

\title{DashReactGrid component}

\description{
DashReactGrid is a wrapper around the ReactGrid component that allows for easy integration with Dash applications.
}

\usage{
dashReactGrid(id=NULL, className=NULL, columns=NULL, data=NULL,
disableVirtualScrolling=NULL, enableColumnSelection=NULL,
enableFillHandle=NULL, enableRangeSelection=NULL,
enableRowSelection=NULL, highlights=NULL, isExtendable=NULL,
persisted_props=NULL, persistence=NULL,
persistence_type=NULL, selectedCell=NULL,
selectedRange=NULL, stickyBottomRows=NULL,
stickyLeftColumns=NULL, stickyRightColumns=NULL,
stickyTopRows=NULL, style=NULL, styleHeader=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks}

\item{className}{Character. }

\item{columns}{Unnamed list. }

\item{data}{Unnamed list. }

\item{disableVirtualScrolling}{Logical. }

\item{enableColumnSelection}{Logical. }

\item{enableFillHandle}{Logical. }

\item{enableRangeSelection}{Logical. }

\item{enableRowSelection}{Logical. }

\item{highlights}{Unnamed list. }

\item{isExtendable}{Logical. }

\item{persisted_props}{Unnamed list. }

\item{persistence}{Logical | character. }

\item{persistence_type}{A value equal to: "local", "session", "memory". }

\item{selectedCell}{Named list. }

\item{selectedRange}{Unnamed list. }

\item{stickyBottomRows}{Numeric. }

\item{stickyLeftColumns}{Numeric. }

\item{stickyRightColumns}{Numeric. }

\item{stickyTopRows}{Numeric. }

\item{style}{Named list. }

\item{styleHeader}{Named list. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

