% Auto-generated: do not edit by hand
\name{dashReactGrid}

\alias{dashReactGrid}

\title{DashReactGrid component}

\description{

}

\usage{
dashReactGrid(id=NULL, columns=NULL, data=NULL,
enableColumnSelection=NULL, enableFillHandle=NULL,
enableRangeSelection=NULL, enableRowSelection=NULL,
highlights=NULL, selectedCell=NULL, stickyBottomRows=NULL,
stickyLeftColumns=NULL, stickyRightColumns=NULL,
stickyTopRows=NULL, style=NULL, styleHeader=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{columns}{Logical | numeric | character | named list | unnamed list. A label that will be printed when this component is rendered.}

\item{data}{Logical | numeric | character | named list | unnamed list. The value displayed in the input.}

\item{enableColumnSelection}{Logical. }

\item{enableFillHandle}{Logical. }

\item{enableRangeSelection}{Logical. }

\item{enableRowSelection}{Logical. }

\item{highlights}{Logical | numeric | character | named list | unnamed list. }

\item{selectedCell}{Logical | numeric | character | named list | unnamed list. }

\item{stickyBottomRows}{Numeric. }

\item{stickyLeftColumns}{Numeric. }

\item{stickyRightColumns}{Numeric. }

\item{stickyTopRows}{Numeric. }

\item{style}{Named list. The style of the container (div)}

\item{styleHeader}{Named list. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

