% Auto-generated: do not edit by hand
\name{dashReactGrid}

\alias{dashReactGrid}

\title{DashReactGrid component}

\description{

}

\usage{
dashReactGrid(id=NULL, className=NULL, columns=NULL, data=NULL,
disableVirtualScrolling=NULL, enableColumnSelection=NULL,
enableFillHandle=NULL, enableRangeSelection=NULL,
enableRowSelection=NULL, highlights=NULL, isExtendable=NULL,
persisted_props=NULL, persistence=NULL,
persistence_type=NULL, selectedCell=NULL,
stickyBottomRows=NULL, stickyLeftColumns=NULL,
stickyRightColumns=NULL, stickyTopRows=NULL, style=NULL,
styleHeader=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{className}{Character. }

\item{columns}{Logical | numeric | character | named list | unnamed list. A label that will be printed when this component is rendered.}

\item{data}{Logical | numeric | character | named list | unnamed list. The value displayed in the input.}

\item{disableVirtualScrolling}{Logical. }

\item{enableColumnSelection}{Logical. }

\item{enableFillHandle}{Logical. }

\item{enableRangeSelection}{Logical. }

\item{enableRowSelection}{Logical. }

\item{highlights}{Logical | numeric | character | named list | unnamed list. }

\item{isExtendable}{Logical. Whether the table automatically adds extra rows}

\item{persisted_props}{List of a value equal to: 'columns', 'data]'s. }

\item{persistence}{Logical | character | numeric. Dash-assigned callback that should be called to report property changes
to Dash, to make them available for callbacks.}

\item{persistence_type}{A value equal to: 'local', 'session', 'memory'. }

\item{selectedCell}{Logical | numeric | character | named list | unnamed list. }

\item{stickyBottomRows}{Numeric. }

\item{stickyLeftColumns}{Numeric. }

\item{stickyRightColumns}{Numeric. }

\item{stickyTopRows}{Numeric. }

\item{style}{Named list. The style of the container (div)}

\item{styleHeader}{Named list. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

